Initial refactor to use Maven to build
