+++ .gitignore
@@ -1,0 +1,5 @@
+*.*~
+.project
+.classpath
+.settings/*
+target/*
