+++ src/main/java/com/sharethis/textrank/Context.java
