+++ src/main/java/com/sharethis/textrank/LanguageEnglish.java
@@ -34,41 +34,48 @@ package com.sharethis.textrank;
 
 import java.io.File;
 
-import opennlp.tools.lang.spanish.PosTagger;
-import opennlp.tools.lang.spanish.SentenceDetector;
-import opennlp.tools.lang.spanish.Tokenizer;
+import opennlp.tools.lang.english.ParserTagger;
+import opennlp.tools.lang.english.SentenceDetector;
+import opennlp.tools.lang.english.Tokenizer;
+import opennlp.tools.sentdetect.SentenceDetectorME;
+import opennlp.tools.util.Sequence;
 
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
 
-import org.tartarus.snowball.ext.spanishStemmer;
+import org.tartarus.snowball.ext.EnglishStemmer;
+
+import spiaotools.SentParDetector;
 
 
 /**
- * Implementation of Spanish-specific tools for natural language
+ * Implementation of English-specific tools for natural language
  * processing.
  *
  * @author paco@sharethis.com
  */
 
 public class
-    LanguageSpanish
+    LanguageEnglish
     extends LanguageModel
 {
     // logging
 
     private final static Log LOG =
-        LogFactory.getLog(LanguageSpanish.class.getName());
+        LogFactory.getLog(LanguageEnglish.class.getName());
 
 
     /**
      * Public definitions.
      */
 
-    public static SentenceDetector splitter_es = null;
-    public static Tokenizer tokenizer_es = null;
-    public static PosTagger tagger_es = null;
-    public static spanishStemmer stemmer_es = null;
+    public static SentParDetector splitter_en = null;
+    /** /
+    public static SentenceDetectorME splitter_en = null;
+    /* */
+    public static Tokenizer tokenizer_en = null;
+    public static ParserTagger tagger_en = null;
+    public static EnglishStemmer stemmer_en = null;
 
 
     /**
@@ -77,10 +84,10 @@ public class
      */
 
     public
-	LanguageSpanish (final String path)
+	LanguageEnglish (final String path)
 	throws Exception
     {
-	if (splitter_es == null) {
+	if (splitter_en == null) {
 	    loadResources(path);
 	}
     }
@@ -94,17 +101,24 @@ public class
 	loadResources (final String path)
 	throws Exception
     {
-	splitter_es =
-	    new SentenceDetector((new File(path, "opennlp/SpanishSent.bin.gz")).getPath());
+	splitter_en = new SentParDetector();
 
-	tokenizer_es =
-	    new Tokenizer((new File(path, "opennlp/SpanishTok.bin.gz")).getPath());
+	/** /
+	splitter_en =
+		new SentenceDetector((new File(path, "opennlp/EnglishSD.bin.gz")).getPath());
+	/* */
 
-	tagger_es =
-	    new PosTagger((new File(path, "opennlp/SpanishPOS.bin.gz")).getPath());
+	tokenizer_en =
+	    new Tokenizer((new File(path, "opennlp/EnglishTok.bin.gz")).getPath());
 
-	stemmer_es =
-	    new spanishStemmer();
+	tagger_en =
+	    new ParserTagger((new File(path, "opennlp/tag.bin.gz")).getPath(),
+			     (new File(path, "opennlp/tagdict")).getPath(),
+			     false
+			     );
+
+	stemmer_en =
+	    new EnglishStemmer();
     }
 
 
@@ -115,7 +129,11 @@ public class
     public String[]
 	splitParagraph (final String text)
     {
-	return splitter_es.sentDetect(text);
+	return splitter_en.markupRawText(2, text).split("\\n");
+
+	/** /
+	return splitter_en.sentDetect(text);
+	/* */
     }
 
 
@@ -126,7 +144,13 @@ public class
     public String[]
 	tokenizeSentence (final String text)
     {
-	return tokenizer_es.tokenize(text);
+	final String[] token_list = tokenizer_en.tokenize(text);
+
+	for (int i = 0; i < token_list.length; i++) {
+	    token_list[i] = token_list[i].replace("\"", "").toLowerCase().trim();
+	}
+
+	return token_list;
     }
 
 
@@ -137,7 +161,17 @@ public class
     public String[]
 	tagTokens (final String[] token_list)
     {
-	return tagger_es.tag(token_list);
+	final Sequence[] sequences = tagger_en.topKSequences(token_list);
+	final String[] tag_list = new String[token_list.length];
+
+	int i = 0;
+
+	for (Object obj : sequences[0].getOutcomes()) {
+	    tag_list[i] = (String) obj;
+	    i++;
+	}
+
+	return tag_list;
     }
 
 
@@ -161,7 +195,7 @@ public class
     public boolean
 	isNoun (final String pos)
     {
-	return pos.startsWith("NC");
+	return pos.startsWith("NN");
     }
 
 
@@ -172,7 +206,7 @@ public class
     public boolean
 	isAdjective (final String pos)
     {
-	return pos.startsWith("AQ");
+	return pos.startsWith("JJ");
     }
 
 
@@ -183,9 +217,9 @@ public class
     public String
 	stemToken (final String token)
     {
-	stemmer_es.setCurrent(token);
-	stemmer_es.stem();
+	stemmer_en.setCurrent(token);
+	stemmer_en.stem();
 
-	return stemmer_es.getCurrent();
+	return stemmer_en.getCurrent();
     }
 }
