+++ src/main/java/com/sharethis/textrank/MetricVector.java
