+++ src/main/java/com/sharethis/textrank/NGram.java
