+++ src/main/java/com/sharethis/textrank/Node.java
