+++ src/main/java/com/sharethis/textrank/NodeValue.java
