+++ src/main/resources/en/wn/adj.exc
