+++ src/main/resources/en/wn/adv.exc
