+++ src/main/resources/en/wn/data.adj
