+++ src/main/resources/en/wn/data.adv
