+++ src/main/resources/en/wn/data.verb
