+++ src/main/resources/en/wn/frames.vrb
