+++ src/main/resources/en/wn/index.adj
