+++ src/main/resources/en/wn/index.verb
