+++ src/main/resources/en/wn/noun.exc
