+++ src/main/resources/en/wn/sentidx.vrb
