+++ src/main/resources/en/wn/sents.vrb
