+++ src/main/resources/en/wn/verb.Framestext
