+++ src/main/resources/en/wn/verb.exc
