+++ src/main/resources/es/opennlp/SpanishTokChunk.bin.gz
