+++ src/test/resources/good.txt
