+++ src/test/resources/kill.txt
